fix(installers): share the artifact-hashing primitives the three installers copied
#996, split out of #681. That issue moved the registry version resolver into a
shared module; this is the set left behind, on a higher-consequence path.

getPlatformString, computeSha256Streaming, hashFile and verifySha256 were
declared separately in terraform-installer.ts, policy-agent-installer.ts and
terraform-docs-installer.ts with byte-identical bodies. They now live in one
tool-integrity.ts, byte-identical across the three src/ directories and
registered in FAMILIES, so check-shared-modules.js fails on any divergence.

Each installer re-exports what it exported before, so no caller and none of the
46 importing test files had to change.

WHY NOTHING REPORTED THIS

Both parity gates were structurally unable to see it. check-shared-modules.js
only compares files named in FAMILIES. check-near-duplicate-modules.js groups
candidates by BASENAME and compares within a group -- these bodies sat in three
differently-named files, so they were never compared to each other, then or now.

That is the general shape worth remembering: the parity machinery can only see
duplication that has ALREADY been extracted into a same-named module. Extracting
is what brings a thing into scope; it is not merely tidying afterwards.

WHAT DELIBERATELY DID NOT MOVE, AND WHY

writeCacheIntegrityMarker and verifyCachedTool are byte-identical across the
three as well, and by rights belong beside these. Moving them broke the build in
a way worth recording rather than working around:

  FAIL: 4 residual instance(s) of the artifact-trust class.
    downloadTerraform()      [CACHE-ADMIT] TRUSTS-CACHE-BLINDLY
    downloadTofu()           [CACHE-ADMIT] TRUSTS-CACHE-BLINDLY
    downloadPolicyAgent()    [CACHE-ADMIT] TRUSTS-CACHE-BLINDLY
    downloadTerraformDocs()  [CACHE-ADMIT] TRUSTS-CACHE-BLINDLY

Nothing had stopped re-verifying. scripts/check-artifact-trust.js resolves a
CACHE-ADMIT verdict WITHIN A SINGLE FILE: its `recordReaders` set is built from
that file's own top-level functions, and the 64-hex marker validator it looks
for must be a module-scope const in the same file. Move the pair out and the
gate can no longer see the re-verification it is there to confirm.

So de-duplicating those two would have blinded the signature that guards them.
They stay put, with the reason recorded at both the module and the FAMILIES
entry rather than left for the next person to rediscover. Teaching that gate to
follow same-directory imports is a change to a security gate and wants its own
commit and its own mutation test; it is filed separately.

Verified: 431 + 264 + 246 tests pass across the three installers;
check-artifact-trust.js still reports "every path by which an artifact becomes
trusted verifies it, discards it on failure, and degrades legibly"; the shared
parity gate now covers 12 families and 20 file comparisons; lint and tsc clean.

Minors bumped on all three so agents re-fetch the changed handlers.

Closes #996.
