+++ Tasks/PolicyAgentInstaller/PolicyAgentInstallerV1/task.json
@@ -14,7 +14,7 @@
     "demands": [],
     "version": {
         "Major": "1",
-        "Minor": "21",
+        "Minor": "25",
         "Patch": "0"
     },
     "minimumAgentVersion": "4.265.1",
