+++ Tasks/TerraformDocsInstaller/TerraformDocsInstallerV1/task.json
@@ -14,7 +14,7 @@
   "demands": [],
   "version": {
     "Major": "1",
-    "Minor": "17",
+    "Minor": "21",
     "Patch": "0"
   },
   "minimumAgentVersion": "4.265.1",
