+++ Tasks/TerraformInstaller/TerraformInstallerV1/task.json
@@ -14,7 +14,7 @@
     "demands": [],
     "version": {
         "Major": "1",
-        "Minor": "245",
+        "Minor": "249",
         "Patch": "0"
     },
     "minimumAgentVersion": "4.265.1",
